requirements.txt
Django==4.2.7
psycopg[binary]==3.2.10
python-decouple==3.8
Pillow==11.0.0
pandas==2.2.3
numpy==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.10

// sudah disesuaikan dengan versi python 3.13.2